Update Propietario Index view to remove hidden ID column
This commit removes the hidden ID column from the Propietario Index view in the ASP.NET MVC application. The ID column was previously displayed as the first column with a class of "d-none" to hide it from view. However, this commit removes that column altogether. The remaining columns display the Propietario's name, last name, and DNI.